fix js build

+++ assembly/instructions/bit.ts
@@ -10,7 +10,7 @@ export const count_set_bits_64: InstructionRun = (r, args, regs) => {
 
 // COUNT_SET_BITS_32
 export const count_set_bits_32: InstructionRun = (r, args, regs) => {
-  regs[reg(args.b)] = portable.popcnt_u32(u32(regs[reg(args.a)]));
+  regs[reg(args.b)] = u64(portable.popcnt_u32(u32(regs[reg(args.a)])));
   return ok(r);
 };
 
@@ -22,7 +22,7 @@ export const leading_zero_bits_64: InstructionRun = (r, args, regs) => {
 
 // LEADING_ZERO_BITS_32
 export const leading_zero_bits_32: InstructionRun = (r, args, regs) => {
-  regs[reg(args.b)] = portable.clz_u32(u32(regs[reg(args.a)]));
+  regs[reg(args.b)] = u64(portable.clz_u32(u32(regs[reg(args.a)])));
   return ok(r);
 };
 
@@ -34,7 +34,7 @@ export const trailing_zero_bits_64: InstructionRun = (r, args, regs) => {
 
 // TRAILING_ZERO_BITS_32
 export const trailing_zero_bits_32: InstructionRun = (r, args, regs) => {
-  regs[reg(args.b)] = portable.ctz_u32(u32(regs[reg(args.a)]));
+  regs[reg(args.b)] = u64(portable.ctz_u32(u32(regs[reg(args.a)])));
   return ok(r);
 };
