+++ assembly/instructions/shift.ts
@@ -24,7 +24,7 @@ export const shlo_r_imm_32: InstructionRun = (r, args, registers) => {
 export const shar_r_imm_32: InstructionRun = (r, args, registers) => {
   const shift = u32(args.c % MAX_SHIFT_32);
   const value = u32SignExtend(u32(registers[reg(args.a)]));
-  registers[reg(args.b)] = value >> i64(shift);
+  registers[reg(args.b)] = u64(i64(value) >> i64(shift));
   return ok(r);
 };
 
@@ -46,14 +46,14 @@ export const shlo_r_imm_alt_32: InstructionRun = (r, args, registers) => {
 export const shar_r_imm_alt_32: InstructionRun = (r, args, registers) => {
   const shift = u32(registers[reg(args.a)] % u64(MAX_SHIFT_32));
   const imm = u32SignExtend(args.c);
-  registers[reg(args.b)] = u32SignExtend(u32(imm >> i64(shift)));
+  registers[reg(args.b)] = u32SignExtend(u32(i64(imm) >> i64(shift)));
   return ok(r);
 };
 
 // SHLO_L_IMM
 export const shlo_l_imm: InstructionRun = (r, args, registers) => {
   const shift = u32(args.c % MAX_SHIFT_64);
-  registers[reg(args.b)] = registers[reg(args.a)] << u64(shift);
+  registers[reg(args.b)] = u64(registers[reg(args.a)] << u64(shift));
   return ok(r);
 };
 
@@ -68,14 +68,14 @@ export const shlo_r_imm: InstructionRun = (r, args, registers) => {
 export const shar_r_imm: InstructionRun = (r, args, registers) => {
   const shift = u32(args.c % MAX_SHIFT_64);
   const value = i64(registers[reg(args.a)]);
-  registers[reg(args.b)] = value >> i64(shift);
+  registers[reg(args.b)] = u64(value >> i64(shift));
   return ok(r);
 };
 
 // SHLO_L_IMM_ALT
 export const shlo_l_imm_alt: InstructionRun = (r, args, registers) => {
   const shift = u32(registers[reg(args.a)] % u64(MAX_SHIFT_64));
-  registers[reg(args.b)] = u32SignExtend(args.c) << i64(shift);
+  registers[reg(args.b)] = u64(u32SignExtend(args.c) << i64(shift));
   return ok(r);
 };
 
@@ -114,14 +114,14 @@ export const shlo_r_32: InstructionRun = (r, args, registers) => {
 export const shar_r_32: InstructionRun = (r, args, registers) => {
   const shift = u32(registers[reg(args.a)] % u64(MAX_SHIFT_32));
   const regValue = u32SignExtend(u32(registers[reg(args.b)]));
-  registers[reg(args.c)] = u32SignExtend(u32(regValue >> i64(shift)));
+  registers[reg(args.c)] = u32SignExtend(u32(i64(regValue) >> i64(shift)));
   return ok(r);
 };
 
 // SHLO_L
 export const shlo_l: InstructionRun = (r, args, registers) => {
   const shift = u32(registers[reg(args.a)] % u64(MAX_SHIFT_64));
-  registers[reg(args.c)] = registers[reg(args.b)] << u64(shift);
+  registers[reg(args.c)] = u64(registers[reg(args.b)] << u64(shift));
   return ok(r);
 };
 
@@ -135,6 +135,6 @@ export const shlo_r: InstructionRun = (r, args, registers) => {
 // SHAR_R
 export const shar_r: InstructionRun = (r, args, registers) => {
   const shift = u32(registers[reg(args.a)] % u64(MAX_SHIFT_64));
-  registers[reg(args.c)] = i64(registers[reg(args.b)]) >> i64(shift);
+  registers[reg(args.c)] = u64(i64(registers[reg(args.b)]) >> i64(shift));
   return ok(r);
 };
